test(e2e): splitted E2E tests for dedicated API. Setup, Ready, Health, Write and Query left in original place.
